Add executable cmake target
Cmake build target cpp-runtime, which builds an executable program file.
It uses the SRC library to optimize build speed.

The main.cpp file hosts a main function. In a CPP program the main
function cannot be tested. Therefore, I move the main out. The source
code will be compiled into a library SRC and is testable. While the
main function will be added in the executable after SRC compiling. It
does only call run() defined in runner.h.

This is not only an optimization, it allows to build catch tests. The
tests have their own main() function and it would result in a name
clash if the gloabal main() would be linked into the test executable.