Update library versions
The catch library did not show enough information about a test failure.
Therefore, I assumed the newer version would print that information. It
does not, but upgrade worked correctly.

The spdlog library throws an exception on construction of the logger. To
ensure, that it is not a bug I upgraded the component. It did not
improve the situation and will look further for the reason.