Type parsing and omission rules
Types are now written out and will be parsed in the linking phase. That
change will simplify writing intermediate file significantly.

Also nearly all options are now optional and do not have to be written
if there is noting to be configured. This shrinks il-code size and
simplifies writing files.